requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
aiohttp==3.9.1
lxml==4.9.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# API & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Task Scheduling
apscheduler==3.10.4

# WebSocket
python-socketio==5.10.0

# Monitoring
prometheus-client==0.19.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# Utils
httpx==0.25.2
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
supervisor==4.2.5